backend/requirements.txt
fastapi==0.115.6
numpy==2.0.2
Pillow==11.1.0
tensorflow==2.18.0
uvicorn==0.34.0
